Move simple_graph fixture conftest.py
